example/comp/requirements.txt
django-debug-toolbar
django-extensions
django-markdown2
django-rosetta
